+++ apps/server/src/utils/arrayUtils.ts
@@ -61,15 +61,3 @@ export const sortArrayByProperty = <T>(arr: T[], property: string): T[] => {
     return a[property] - b[property];
   });
 };
-
-/**
- * @description Creates a nested object with keys from an array and assigns the `value` to the last key
- * @param {array} arr - array to be nested
- * @param {string} value - value to assign
- * @returns {object | null} nested object or null if no object was created
- */
-export const nestedObjectFromArray = (arr: string[], value?: unknown): object | null => {
-  const obj = arr.reduceRight((result, key) => ({ [key]: result }), value);
-  if (typeof obj === 'object') return obj;
-  return null;
-};
